requirements.txt
python-telegram-bot>=20.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2024.1